+++ backend/src/tickets/tickets.router.js
@@ -1,5 +1,5 @@
 const router = require('express').Router({ mergeParams: true })
-const controller = require('./cards.controller')
+const controller = require('./tickets.controller')
 const methodNotAllowed = require('../errors/methodNotAllowed')
 
 router.route('/').get(controller.list).all(methodNotAllowed)
